+++ src/Core/Model/Cart/ScoreShippingRateInputDraft.php
@@ -13,8 +13,8 @@ use Commercetools\Core\Model\Common\Context;
  * @link http://dev.commercetools.com/http-api-projects-carts.html#scoreshippingrateinputdraft
  * @method string getType()
  * @method ScoreShippingRateInputDraft setType(string $type = null)
- * @method float getScore()
- * @method ScoreShippingRateInputDraft setScore(float $score = null)
+ * @method int getScore()
+ * @method ScoreShippingRateInputDraft setScore(int $score = null)
  */
 class ScoreShippingRateInputDraft extends ShippingRateInputDraft
 {
@@ -22,12 +22,12 @@ class ScoreShippingRateInputDraft extends ShippingRateInputDraft
     {
         return [
             'type' => [static::TYPE => 'string'],
-            'score' => [static::TYPE => 'float'],
+            'score' => [static::TYPE => 'int'],
         ];
     }
 
     /**
-     * @param $score
+     * @param int $score
      * @param Context|callable $context
      * @return ScoreShippingRateInputDraft
      */
